requirements.txt
fastapi==0.110.3
uvicorn==0.30.1
pydantic==1.10.13
python-dotenv==1.0.1
pypdf==4.2.0
pandas==2.2.2
# Optional: use CrewAI if available; app will gracefully degrade if it's not installed
crewai==0.130.0
crewai-tools==0.47.1
redis==5.0.7
celery==5.4.0
SQLAlchemy==2.0.32
